requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.12.1
Automat==22.10.0
billiard==3.6.4.0
boto3==1.17.94
botocore==1.20.94
celery==5.2.7
certifi==2022.6.15.1
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.1
daphne==4.0.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.5
Django==4.1.1
django-allauth==0.51.0
django-common-helpers==0.9.2
django-cors-headers==3.13.0
django-crispy-forms==1.12.0
django-cron==0.6.0
django-filter==2.4.0
django-heroku==0.3.1
django-resized==1.0.2
django-storages==1.11.1
django-user-agents==0.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
frozenlist==1.3.1
geoip2==4.6.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
jmespath==0.10.0
kombu==5.2.4
Markdown==3.3.3
maxminddb==2.2.0
msgpack==1.0.4
multidict==6.0.2
oauthlib==3.2.1
Pillow==9.2.0
prompt-toolkit==3.0.31
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==23.0.0
python-dateutil==2.8.1
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==5.4.1
redis==4.4.1
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.4.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing-extensions==3.10.0.0
tzdata==2022.2
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
yarl==1.8.1
zope.interface==5.5.2
